chore(tests): prop diffing & updates, harden reconstruct, isolate reconciler cases
