+++ packages/fiber/tests/index.test.tsx
@@ -1,0 +1,220 @@
+import * as React from 'react'
+import * as THREE from 'three'
+import { ReconcilerRoot, createRoot, act, useFrame, useThree, createPortal, RootState } from '../src/index'
+import { UseBoundStore } from 'zustand'
+import { privateKeys } from '../src/core/store'
+
+let root: ReconcilerRoot<HTMLCanvasElement> = null!
+
+beforeEach(() => (root = createRoot(document.createElement('canvas'))))
+afterEach(async () => act(async () => root.unmount()))
+
+describe('createRoot', () => {
+  it('should return a Zustand store', async () => {
+    const store = await act(async () => root.render(null))
+    expect(() => store.getState()).not.toThrow()
+  })
+
+  it('will make an Orthographic Camera & set the position', async () => {
+    const store = await act(async () =>
+      root.configure({ orthographic: true, camera: { position: [0, 0, 5] } }).render(<group />),
+    )
+    const { camera } = store.getState()
+
+    expect(camera).toBeInstanceOf(THREE.OrthographicCamera)
+    expect(camera.position.z).toEqual(5)
+  })
+
+  it('should handle an performance changing functions', async () => {
+    let state: UseBoundStore<RootState> = null!
+    await act(async () => {
+      state = root.configure({ dpr: [1, 2], performance: { min: 0.2 } }).render(<group />)
+    })
+
+    expect(state.getState().viewport.initialDpr).toEqual(window.devicePixelRatio)
+    expect(state.getState().performance.min).toEqual(0.2)
+    expect(state.getState().performance.current).toEqual(1)
+
+    await act(async () => {
+      state.getState().setDpr(0.1)
+    })
+
+    expect(state.getState().viewport.dpr).toEqual(0.1)
+
+    jest.useFakeTimers()
+
+    await act(async () => {
+      state.getState().performance.regress()
+      jest.advanceTimersByTime(100)
+    })
+
+    expect(state.getState().performance.current).toEqual(0.2)
+
+    await act(async () => {
+      jest.advanceTimersByTime(200)
+    })
+
+    expect(state.getState().performance.current).toEqual(1)
+
+    jest.useRealTimers()
+  })
+
+  it('should set PCFSoftShadowMap as the default shadow map', async () => {
+    const store = await act(async () => root.configure({ shadows: true }).render(<group />))
+    const { gl } = store.getState()
+
+    expect(gl.shadowMap.type).toBe(THREE.PCFSoftShadowMap)
+  })
+
+  it('should set tonemapping to ACESFilmicToneMapping and outputEncoding to sRGBEncoding if linear is false', async () => {
+    const store = await act(async () => root.configure({ linear: false }).render(<group />))
+    const { gl } = store.getState()
+
+    expect(gl.toneMapping).toBe(THREE.ACESFilmicToneMapping)
+    expect(gl.outputEncoding).toBe(THREE.sRGBEncoding)
+  })
+
+  it('should toggle render mode in xr', async () => {
+    let state: RootState = null!
+
+    await act(async () => {
+      state = root.render(<group />).getState()
+      state.gl.xr.isPresenting = true
+      state.gl.xr.dispatchEvent({ type: 'sessionstart' })
+    })
+
+    expect(state.gl.xr.enabled).toEqual(true)
+
+    await act(async () => {
+      state.gl.xr.isPresenting = false
+      state.gl.xr.dispatchEvent({ type: 'sessionend' })
+    })
+
+    expect(state.gl.xr.enabled).toEqual(false)
+  })
+
+  it('should respect frameloop="never" in xr', async () => {
+    let respected = true
+
+    const Test = () => useFrame(() => (respected = false))
+
+    await act(async () => {
+      const state = root
+        .configure({ frameloop: 'never' })
+        .render(<Test />)
+        .getState()
+      state.gl.xr.isPresenting = true
+      state.gl.xr.dispatchEvent({ type: 'sessionstart' })
+    })
+
+    expect(respected).toEqual(true)
+  })
+
+  it('should set renderer props via gl prop', async () => {
+    const store = await act(async () => root.configure({ gl: { physicallyCorrectLights: true } }).render(<group />))
+    const { gl } = store.getState()
+
+    expect(gl.physicallyCorrectLights).toBe(true)
+  })
+
+  it('should set a renderer via gl callback', async () => {
+    class Renderer extends THREE.WebGLRenderer {}
+
+    const store = await act(async () => root.configure({ gl: (canvas) => new Renderer({ canvas }) }).render(<group />))
+    const { gl } = store.getState()
+
+    expect(gl instanceof Renderer).toBe(true)
+  })
+
+  it('should respect color management preferences via gl', async () => {
+    const store = await act(async () =>
+      root
+        .configure({ gl: { outputEncoding: THREE.LinearEncoding, toneMapping: THREE.NoToneMapping } })
+        .render(<group />),
+    )
+    const { gl } = store.getState()
+
+    expect(gl.outputEncoding).toBe(THREE.LinearEncoding)
+    expect(gl.toneMapping).toBe(THREE.NoToneMapping)
+
+    await act(async () => root.configure({ flat: true, linear: true }).render(<group />))
+    expect(gl.outputEncoding).toBe(THREE.LinearEncoding)
+    expect(gl.toneMapping).toBe(THREE.NoToneMapping)
+  })
+
+  it('should respect legacy prop', async () => {
+    await act(async () => root.configure({ legacy: true }).render(<group />))
+    expect((THREE as any).ColorManagement.legacyMode).toBe(true)
+
+    await act(async () => root.configure({ legacy: false }).render(<group />))
+    expect((THREE as any).ColorManagement.legacyMode).toBe(false)
+  })
+})
+
+describe('createPortal', () => {
+  it('should create a state enclave', async () => {
+    const scene = new THREE.Scene()
+
+    let state: RootState = null!
+    let portalState: RootState = null!
+
+    const Normal = () => {
+      const three = useThree()
+      state = three
+
+      return <group />
+    }
+
+    const Portal = () => {
+      const three = useThree()
+      portalState = three
+
+      return <group />
+    }
+
+    await act(async () => {
+      root.render(
+        <>
+          <Normal />
+          {createPortal(<Portal />, scene, { scene })}
+        </>,
+      )
+    })
+
+    // Renders into portal target
+    expect(scene.children.length).not.toBe(0)
+
+    // Creates an isolated state enclave
+    expect(state.scene).not.toBe(scene)
+    expect(portalState.scene).toBe(scene)
+
+    // Preserves internal keys
+    const overwrittenKeys = ['get', 'set', 'events', 'size', 'viewport']
+    const respectedKeys = privateKeys.filter((key) => overwrittenKeys.includes(key) || state[key] === portalState[key])
+    expect(respectedKeys).toStrictEqual(privateKeys)
+  })
+
+  it('should handle unmounted containers', async () => {
+    let groupHandle!: THREE.Group | null
+    function Test(props: any) {
+      const [group, setGroup] = React.useState(null)
+      groupHandle = group
+
+      return (
+        <group {...props} ref={setGroup}>
+          {group && createPortal(<mesh />, group)}
+        </group>
+      )
+    }
+
+    await act(async () => root.render(<Test key={0} />))
+
+    expect(groupHandle).toBeDefined()
+    const prevUUID = groupHandle!.uuid
+
+    await act(async () => root.render(<Test key={1} />))
+
+    expect(groupHandle).toBeDefined()
+    expect(prevUUID).not.toBe(groupHandle!.uuid)
+  })
+})
