+++ packages/fiber/tests/renderer.test.tsx
@@ -1,18 +1,6 @@
 import * as React from 'react'
 import * as THREE from 'three'
-import {
-  ReconcilerRoot,
-  createRoot,
-  act,
-  useFrame,
-  extend,
-  ThreeElement,
-  useThree,
-  createPortal,
-  RootState,
-} from '../src/index'
-import { UseBoundStore } from 'zustand'
-import { privateKeys } from '../src/core/store'
+import { ReconcilerRoot, createRoot, act, extend, ThreeElement } from '../src/index'
 import { suspend } from 'suspend-react'
 
 class CustomElement extends THREE.Object3D {}
@@ -26,6 +14,22 @@ declare module '@react-three/fiber' {
 extend({ CustomElement })
 
 type ComponentMesh = THREE.Mesh<THREE.BoxBufferGeometry, THREE.MeshBasicMaterial>
+
+const expectToThrow = async (callback: () => any) => {
+  const error = console.error
+  console.error = jest.fn()
+
+  let thrown = false
+  try {
+    await callback()
+  } catch (_) {
+    thrown = true
+  }
+
+  expect(thrown).toBe(true)
+  expect(console.error).toBeCalled()
+  console.error = error
+}
 
 describe('renderer', () => {
   let root: ReconcilerRoot<HTMLCanvasElement> = null!
@@ -163,6 +167,154 @@ describe('renderer', () => {
     expect(lifecycle).toStrictEqual(['attach', 'mount'])
   })
 
+  it('should update props reactively', async () => {
+    const store = await act(async () => root.render(<group />))
+    const { scene } = store.getState()
+    const group = scene.children[0] as THREE.Group
+
+    // Initial
+    expect(group.name).toBe(new THREE.Group().name)
+
+    // Set
+    await act(async () => root.render(<group name="one" />))
+    expect(group.name).toBe('one')
+
+    // Update
+    await act(async () => root.render(<group name="two" />))
+    expect(group.name).toBe('two')
+
+    // Unset
+    await act(async () => root.render(<group />))
+    expect(group.name).toBe(new THREE.Group().name)
+  })
+
+  it('should handle event props reactively', async () => {
+    const store = await act(async () => root.render(<mesh />))
+    const { scene, internal } = store.getState()
+    const mesh = scene.children[0] as ComponentMesh
+    mesh.name = 'current'
+
+    // Initial
+    expect(internal.interaction.length).toBe(0)
+
+    // Set
+    await act(async () => root.render(<mesh onClick={() => void 0} />))
+    expect(internal.interaction.length).toBe(1)
+    expect(internal.interaction).toStrictEqual([mesh])
+
+    // Update
+    await act(async () => root.render(<mesh onPointerOver={() => void 0} />))
+    expect(internal.interaction.length).toBe(1)
+    expect(internal.interaction).toStrictEqual([mesh])
+
+    // Unset
+    await act(async () => root.render(<mesh />))
+    expect(internal.interaction.length).toBe(0)
+  })
+
+  it('should handle the args prop reactively', async () => {
+    const ref = React.createRef<ComponentMesh>()
+    const child = React.createRef<THREE.Object3D>()
+    const attachedChild = React.createRef<THREE.Object3D>()
+
+    const Test = (props: JSX.IntrinsicElements['mesh']) => (
+      <mesh {...props} ref={ref}>
+        <object3D ref={child} />
+        <object3D ref={attachedChild} attach="userData-attach" />
+      </mesh>
+    )
+
+    // Initial
+    await act(async () => root.render(<Test />))
+    expect(ref.current!.geometry).toBeInstanceOf(THREE.BufferGeometry)
+    expect(ref.current!.geometry).not.toBeInstanceOf(THREE.BoxGeometry)
+    expect(ref.current!.material).toBeInstanceOf(THREE.Material)
+    expect(ref.current!.material).not.toBeInstanceOf(THREE.MeshStandardMaterial)
+    expect(ref.current!.children[0]).toBe(child.current)
+    expect(ref.current!.userData.attach).toBe(attachedChild.current)
+
+    // Throw on non-array value
+    await expectToThrow(
+      // @ts-expect-error
+      async () => await act(async () => root.render(<Test args={{}} />)),
+    )
+
+    // Set
+    const geometry1 = new THREE.BoxGeometry()
+    const material1 = new THREE.MeshStandardMaterial()
+    await act(async () => root.render(<Test args={[geometry1, material1]} />))
+    expect(ref.current!.geometry).toBe(geometry1)
+    expect(ref.current!.material).toBe(material1)
+    expect(ref.current!.children[0]).toBe(child.current)
+    expect(ref.current!.userData.attach).toBe(attachedChild.current)
+
+    // Update
+    const geometry2 = new THREE.BoxGeometry()
+    const material2 = new THREE.MeshStandardMaterial()
+    await act(async () => root.render(<Test args={[geometry2, material2]} />))
+    expect(ref.current!.geometry).toBe(geometry2)
+    expect(ref.current!.material).toBe(material2)
+    expect(ref.current!.children[0]).toBe(child.current)
+    expect(ref.current!.userData.attach).toBe(attachedChild.current)
+
+    // Unset
+    await act(async () => root.render(<Test />))
+    expect(ref.current!.geometry).toBeInstanceOf(THREE.BufferGeometry)
+    expect(ref.current!.geometry).not.toBeInstanceOf(THREE.BoxGeometry)
+    expect(ref.current!.material).toBeInstanceOf(THREE.Material)
+    expect(ref.current!.material).not.toBeInstanceOf(THREE.MeshStandardMaterial)
+    expect(ref.current!.children[0]).toBe(child.current)
+    expect(ref.current!.userData.attach).toBe(attachedChild.current)
+  })
+
+  it('should handle the object prop reactively', async () => {
+    const ref = React.createRef<THREE.Object3D>()
+    const child = React.createRef<THREE.Object3D>()
+    const attachedChild = React.createRef<THREE.Object3D>()
+
+    const Test = (props: JSX.IntrinsicElements['primitive']) => (
+      <primitive {...props} ref={ref}>
+        <object3D ref={child} />
+        <object3D ref={attachedChild} attach="userData-attach" />
+      </primitive>
+    )
+
+    const object1 = new THREE.Object3D()
+    const child1 = new THREE.Object3D()
+    object1.add(child1)
+
+    const object2 = new THREE.Object3D()
+    const child2 = new THREE.Object3D()
+    object2.add(child2)
+
+    // Initial
+    await act(async () => root.render(<Test object={object1} />))
+    expect(ref.current).toBe(object1)
+    expect(ref.current!.children[0]).toBe(child1)
+    expect(ref.current!.children[1]).toBe(child.current)
+    expect(ref.current!.userData.attach).toBe(attachedChild.current)
+
+    // Throw on undefined
+    await expectToThrow(
+      // @ts-expect-error
+      async () => await act(async () => root.render(<Test object={undefined} />)),
+    )
+
+    // Update
+    await act(async () => root.render(<Test object={object2} />))
+    expect(ref.current).toBe(object2)
+    expect(ref.current!.children[0]).toBe(child2)
+    expect(ref.current!.children[1]).toBe(child.current)
+    expect(ref.current!.userData.attach).toBe(attachedChild.current)
+
+    // Revert
+    await act(async () => root.render(<Test object={object1} />))
+    expect(ref.current).toBe(object1)
+    expect(ref.current!.children[0]).toBe(child1)
+    expect(ref.current!.children[1]).toBe(child.current)
+    expect(ref.current!.userData.attach).toBe(attachedChild.current)
+  })
+
   it('should handle unmount', async () => {
     const dispose = jest.fn()
     const childDispose = jest.fn()
@@ -184,7 +336,8 @@ describe('renderer', () => {
         ref={(self: any) => {
           if (!self) return
           self.dispose = dispose
-        }}>
+        }}
+        onClick={() => void 0}>
         <object3D
           ref={(self: any) => {
             if (!self) return
@@ -219,10 +372,12 @@ describe('renderer', () => {
     const store = await act(async () => root.render(<Test />))
     await act(async () => root.render(null))
 
-    const { scene } = store.getState()
+    const { scene, internal } = store.getState()
 
     // Cleans up scene-graph
     expect(scene.children.length).toBe(0)
+    // Removes events
+    expect(internal.interaction.length).toBe(0)
     // Calls dispose on top-level instance
     expect(dispose).toBeCalled()
     // Also disposes of children
@@ -239,352 +394,6 @@ describe('renderer', () => {
     // Only disposes of declarative primitive children
     expect(objectExternal.dispose).not.toBeCalled()
     expect(disposeDeclarativePrimitive).toBeCalled()
-  })
-
-  it('updates types & names', async () => {
-    const store = await act(async () =>
-      root.render(
-        <mesh>
-          <meshBasicMaterial name="basicMat">
-            <color attach="color" args={[0, 0, 0]} />
-          </meshBasicMaterial>
-        </mesh>,
-      ),
-    )
-    const { scene } = store.getState()
-
-    const basic = scene.children[0] as ComponentMesh
-    expect(basic.material).toBeInstanceOf(THREE.MeshBasicMaterial)
-    expect(basic.material.name).toBe('basicMat')
-    expect(basic.material.color.toArray()).toStrictEqual([0, 0, 0])
-
-    await act(async () =>
-      root.render(
-        <mesh>
-          <meshStandardMaterial name="standardMat">
-            <color attach="color" args={[255, 255, 255]} />
-          </meshStandardMaterial>
-        </mesh>,
-      ),
-    )
-
-    const standard = scene.children[0] as ComponentMesh
-    expect(standard.material).toBeInstanceOf(THREE.MeshStandardMaterial)
-    expect(standard.material.name).toBe('standardMat')
-    expect(standard.material.color.toArray()).toStrictEqual([255, 255, 255])
-  })
-
-  it('will mount/unmount event handlers correctly', async () => {
-    let state: RootState = null!
-    let mounted = false
-    let attachEvents = false
-
-    const EventfulComponent = () => (mounted ? <group onClick={attachEvents ? () => void 0 : undefined} /> : null)
-
-    // Test initial mount without events
-    mounted = true
-    await act(async () => {
-      state = root.render(<EventfulComponent />).getState()
-    })
-    expect(state.internal.interaction.length).toBe(0)
-
-    // Test initial mount with events
-    attachEvents = true
-    await act(async () => {
-      state = root.render(<EventfulComponent />).getState()
-    })
-    expect(state.internal.interaction.length).not.toBe(0)
-
-    // Test events update
-    attachEvents = false
-    await act(async () => {
-      state = root.render(<EventfulComponent />).getState()
-    })
-    expect(state.internal.interaction.length).toBe(0)
-
-    attachEvents = true
-    await act(async () => {
-      state = root.render(<EventfulComponent />).getState()
-    })
-    expect(state.internal.interaction.length).not.toBe(0)
-
-    // Test unmount with events
-    mounted = false
-    await act(async () => {
-      state = root.render(<EventfulComponent />).getState()
-    })
-    expect(state.internal.interaction.length).toBe(0)
-  })
-
-  it('will create an identical instance when reconstructing', async () => {
-    const instances: { uuid: string; parentUUID?: string; childUUID?: string }[] = []
-
-    const object1 = new THREE.Group()
-    const object2 = new THREE.Group()
-
-    const Test = ({ first }: { first?: boolean }) => (
-      <primitive object={first ? object1 : object2} onPointerOver={() => null}>
-        <group />
-      </primitive>
-    )
-
-    const store = await act(async () => root.render(<Test first />))
-    const { scene, internal } = store.getState()
-
-    instances.push({
-      uuid: scene.children[0].uuid,
-      parentUUID: scene.children[0].parent?.uuid,
-      childUUID: scene.children[0].children[0]?.uuid,
-    })
-    expect(scene.children[0]).toBe(object1)
-
-    await act(async () => root.render(<Test />))
-
-    instances.push({
-      uuid: scene.children[0].uuid,
-      parentUUID: scene.children[0].parent?.uuid,
-      childUUID: scene.children[0].children[0]?.uuid,
-    })
-
-    const [oldInstance, newInstance] = instances
-
-    // Swapped to new instance
-    expect(scene.children[0]).toBe(object2)
-
-    // Preserves scene hierarchy
-    expect(oldInstance.parentUUID).toBe(newInstance.parentUUID)
-    expect(oldInstance.childUUID).toBe(newInstance.childUUID)
-
-    // Rebinds events
-    expect(internal.interaction.length).not.toBe(0)
-  })
-
-  it('can swap primitives', async () => {
-    const o1 = new THREE.Group()
-    o1.add(new THREE.Group())
-    const o2 = new THREE.Group()
-
-    const Test = ({ n }: { n: number }) => (
-      <primitive object={n === 1 ? o1 : o2}>
-        <group attach="test" />
-      </primitive>
-    )
-
-    const store = await act(async () => root.render(<Test n={1} />))
-    const { scene } = store.getState()
-
-    // Initial object is added with children and attachments
-    expect(scene.children[0]).toBe(o1)
-    expect(scene.children[0].children.length).toBe(1)
-    expect((scene.children[0] as any).test).toBeInstanceOf(THREE.Group)
-
-    await act(async () => root.render(<Test n={2} />))
-
-    // Swapped to object 2, does not copy old children, copies attachments
-    expect(scene.children[0]).toBe(o2)
-    expect(scene.children[0].children.length).toBe(0)
-    expect((scene.children[0] as any).test).toBeInstanceOf(THREE.Group)
-  })
-
-  it('will make an Orthographic Camera & set the position', async () => {
-    const store = await act(async () =>
-      root.configure({ orthographic: true, camera: { position: [0, 0, 5] } }).render(<group />),
-    )
-    const { camera } = store.getState()
-
-    expect(camera).toBeInstanceOf(THREE.OrthographicCamera)
-    expect(camera.position.z).toEqual(5)
-  })
-
-  it('should handle an performance changing functions', async () => {
-    let state: UseBoundStore<RootState> = null!
-    await act(async () => {
-      state = root.configure({ dpr: [1, 2], performance: { min: 0.2 } }).render(<group />)
-    })
-
-    expect(state.getState().viewport.initialDpr).toEqual(window.devicePixelRatio)
-    expect(state.getState().performance.min).toEqual(0.2)
-    expect(state.getState().performance.current).toEqual(1)
-
-    await act(async () => {
-      state.getState().setDpr(0.1)
-    })
-
-    expect(state.getState().viewport.dpr).toEqual(0.1)
-
-    jest.useFakeTimers()
-
-    await act(async () => {
-      state.getState().performance.regress()
-      jest.advanceTimersByTime(100)
-    })
-
-    expect(state.getState().performance.current).toEqual(0.2)
-
-    await act(async () => {
-      jest.advanceTimersByTime(200)
-    })
-
-    expect(state.getState().performance.current).toEqual(1)
-
-    jest.useRealTimers()
-  })
-
-  it('should set PCFSoftShadowMap as the default shadow map', async () => {
-    const store = await act(async () => root.configure({ shadows: true }).render(<group />))
-    const { gl } = store.getState()
-
-    expect(gl.shadowMap.type).toBe(THREE.PCFSoftShadowMap)
-  })
-
-  it('should set tonemapping to ACESFilmicToneMapping and outputEncoding to sRGBEncoding if linear is false', async () => {
-    const store = await act(async () => root.configure({ linear: false }).render(<group />))
-    const { gl } = store.getState()
-
-    expect(gl.toneMapping).toBe(THREE.ACESFilmicToneMapping)
-    expect(gl.outputEncoding).toBe(THREE.sRGBEncoding)
-  })
-
-  it('should toggle render mode in xr', async () => {
-    let state: RootState = null!
-
-    await act(async () => {
-      state = root.render(<group />).getState()
-      state.gl.xr.isPresenting = true
-      state.gl.xr.dispatchEvent({ type: 'sessionstart' })
-    })
-
-    expect(state.gl.xr.enabled).toEqual(true)
-
-    await act(async () => {
-      state.gl.xr.isPresenting = false
-      state.gl.xr.dispatchEvent({ type: 'sessionend' })
-    })
-
-    expect(state.gl.xr.enabled).toEqual(false)
-  })
-
-  it('should respect frameloop="never" in xr', async () => {
-    let respected = true
-
-    const Test = () => useFrame(() => (respected = false))
-
-    await act(async () => {
-      const state = root
-        .configure({ frameloop: 'never' })
-        .render(<Test />)
-        .getState()
-      state.gl.xr.isPresenting = true
-      state.gl.xr.dispatchEvent({ type: 'sessionstart' })
-    })
-
-    expect(respected).toEqual(true)
-  })
-
-  it('should set renderer props via gl prop', async () => {
-    const store = await act(async () => root.configure({ gl: { physicallyCorrectLights: true } }).render(<group />))
-    const { gl } = store.getState()
-
-    expect(gl.physicallyCorrectLights).toBe(true)
-  })
-
-  it('should set a renderer via gl callback', async () => {
-    class Renderer extends THREE.WebGLRenderer {}
-
-    const store = await act(async () => root.configure({ gl: (canvas) => new Renderer({ canvas }) }).render(<group />))
-    const { gl } = store.getState()
-
-    expect(gl instanceof Renderer).toBe(true)
-  })
-
-  it('should respect color management preferences via gl', async () => {
-    const store = await act(async () =>
-      root
-        .configure({ gl: { outputEncoding: THREE.LinearEncoding, toneMapping: THREE.NoToneMapping } })
-        .render(<group />),
-    )
-    const { gl } = store.getState()
-
-    expect(gl.outputEncoding).toBe(THREE.LinearEncoding)
-    expect(gl.toneMapping).toBe(THREE.NoToneMapping)
-
-    await act(async () => root.configure({ flat: true, linear: true }).render(<group />))
-    expect(gl.outputEncoding).toBe(THREE.LinearEncoding)
-    expect(gl.toneMapping).toBe(THREE.NoToneMapping)
-  })
-
-  it('should respect legacy prop', async () => {
-    await act(async () => root.configure({ legacy: true }).render(<group />))
-    expect((THREE as any).ColorManagement.legacyMode).toBe(true)
-
-    await act(async () => root.configure({ legacy: false }).render(<group />))
-    expect((THREE as any).ColorManagement.legacyMode).toBe(false)
-  })
-
-  it('can handle createPortal', async () => {
-    const scene = new THREE.Scene()
-
-    let state: RootState = null!
-    let portalState: RootState = null!
-
-    const Normal = () => {
-      const three = useThree()
-      state = three
-
-      return <group />
-    }
-
-    const Portal = () => {
-      const three = useThree()
-      portalState = three
-
-      return <group />
-    }
-
-    await act(async () => {
-      root.render(
-        <>
-          <Normal />
-          {createPortal(<Portal />, scene, { scene })}
-        </>,
-      )
-    })
-
-    // Renders into portal target
-    expect(scene.children.length).not.toBe(0)
-
-    // Creates an isolated state enclave
-    expect(state.scene).not.toBe(scene)
-    expect(portalState.scene).toBe(scene)
-
-    // Preserves internal keys
-    const overwrittenKeys = ['get', 'set', 'events', 'size', 'viewport']
-    const respectedKeys = privateKeys.filter((key) => overwrittenKeys.includes(key) || state[key] === portalState[key])
-    expect(respectedKeys).toStrictEqual(privateKeys)
-  })
-
-  it('can handle createPortal on unmounted container', async () => {
-    let groupHandle!: THREE.Group | null
-    function Test(props: any) {
-      const [group, setGroup] = React.useState(null)
-      groupHandle = group
-
-      return (
-        <group {...props} ref={setGroup}>
-          {group && createPortal(<mesh />, group)}
-        </group>
-      )
-    }
-
-    await act(async () => root.render(<Test key={0} />))
-
-    expect(groupHandle).toBeDefined()
-    const prevUUID = groupHandle!.uuid
-
-    await act(async () => root.render(<Test key={1} />))
-
-    expect(groupHandle).toBeDefined()
-    expect(prevUUID).not.toBe(groupHandle!.uuid)
   })
 
   it('should gracefully handle text', async () => {
@@ -618,7 +427,7 @@ describe('renderer', () => {
           <group
             name="child"
             ref={(self) => void (lastMounted = self?.uuid)}
-            attach={(parent, self) => {
+            attach={(_, self) => {
               calls.push('attach')
               lastAttached = self.uuid
               return () => calls.push('detach')
